[ModView] 64-bit fix for ulong type
Previously, ulong was typedef'd to unsigned long which is 64-bits on
some 64-bit systems. This commit replaces unsigned long with unsigned
int. While still not strictly correct, it was the easiest way to fix it.

+++ tools/ModView/png/png.h
@@ -36,7 +36,7 @@
 
 typedef unsigned char	byte;
 typedef unsigned short	word;
-typedef unsigned long	ulong;
+typedef unsigned int	ulong;
 
 #pragma pack(push)
 #pragma pack(1)
